move module to make room for the password change views
